+++ packages/dash-table/src/dash-table/components/Table/props.ts
@@ -7,11 +7,13 @@ export type Columns = IColumn[];
 export type Dataframe = Datum[];
 export type Datum =  IDatumObject | any;
 export type Filtering = 'fe' | 'be' | boolean;
+export type Indices = number[];
 export type Navigation = 'page';
 export type RowSelection = 'single' | 'multi' | false;
 export type SelectedCells = CellCoordinates[];
 export type SetProps = (...args: any[]) => void;
 export type Sorting = 'fe' | 'be' | boolean;
+export type SortingType = 'multi' | 'single';
 export type Virtualization = 'fe' | 'be' | boolean;
 
 interface IColumn {
@@ -62,6 +64,7 @@ interface IProps {
     setProps?: SetProps;
     sorting?: Sorting;
     sorting_settings?: SortSettings;
+    sorting_type?: SortingType;
     table_style?: { selector: string, rule: string }[];
     virtualization?: Virtualization;
     virtualization_settings?: IVirtualizationSettings;
@@ -90,9 +93,10 @@ interface IDefaultProps {
     selected_rows: number[];
     sorting: Sorting;
     sorting_settings: SortSettings;
+    sorting_type: SortingType;
     table_style: { selector: string, rule: string }[];
-    virtual_dataframe: any[];
-    virtual_dataframe_indices: number[];
+    virtual_dataframe: Dataframe;
+    virtual_dataframe_indices: Indices;
     virtualization: Virtualization;
     virtualization_settings: IVirtualizationSettings;
 }
@@ -111,6 +115,7 @@ export interface ICellFactoryOptions {
     column_conditional_styles: any[];
     column_static_dropdown: any;
     column_static_style: any;
+    dataframe: Dataframe;
     editable: boolean;
     id: string;
     is_focused?: boolean;
@@ -121,5 +126,7 @@ export interface ICellFactoryOptions {
     selected_cell: SelectedCells;
     selected_rows: number[];
     setProps: SetProps;
+    virtual_dataframe: Dataframe;
+    virtual_dataframe_indices: Indices;
     virtualizer: AbstractVirtualizationStrategy;
 }