+++ packages/dash-table/CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+## RC6 - Sorting props
+
+- Additional sorting_type prop that can take value 'multi' or 'single'
+    This prop defines whether the user can sort based on multiple columns or can only sort by one column at a time. The default value is 'single'.